# Move CRD reader to istio/client-go