… go.sum
… pilot/pkg/config/kube/crd/codegen/types.go
… pilot/pkg/config/kube/crd/codegen/types.go.tmpl
… pilot/pkg/config/kube/crd/config.go
… pilot/pkg/config/kube/crd/config_test.go
… pilot/pkg/config/kube/crd/controller/client.go
… pilot/pkg/config/kube/crd/controller/controller.go
… pilot/pkg/config/kube/crd/types.gen.go
… pilot/pkg/config/kube/crdclient/client.go
… pilot/pkg/config/kube/crdclient/client_test.go
… pilot/pkg/config/kube/crdclient/gen/types.go.tmpl
… pilot/pkg/config/kube/crdclient/ledger.go
… pilot/pkg/config/kube/crdclient/metrics.go
… pilot/pkg/config/kube/crdclient/types.gen.go
… pilot/pkg/config/kube/gateway/controller_test.go
… pilot/pkg/config/kube/gateway/conversion.go
… pilot/pkg/status/report.go
… pilot/test/mock/config.go
… pkg/config/schema/collection/schemas.go
… pkg/config/schema/collections/collections.gen.go
… pkg/config/schema/collections/staticinit.gen.go
… pkg/config/schema/metadata.gen.go
… pkg/config/schema/metadata.yaml
… pkg/config/schema/resource/schema.go
… pkg/config/schema/snapshots/staticinit.gen.go
… pkg/config/schema/staticinit.gen.go