This moves Pilot's internal CRD reading code to use istio/client-go

Benefits/motivation:
* perf is *much* better due to less (un)marshalling:

```
$ benchstat /tmp/{old,new}
name        old time/op    new time/op    delta
CRD/list-8    29.9µs ± 0%     1.5µs ± 0%   ~     (p=1.000 n=1+1)
CRD/get-8     5.24ms ± 0%    0.08ms ± 0%   ~     (p=1.000 n=1+1)

name        old alloc/op   new alloc/op   delta
CRD/list-8    9.03kB ± 0%    0.32kB ± 0%   ~     (p=1.000 n=1+1)
CRD/get-8     1.90MB ± 0%    0.09MB ± 0%   ~     (p=1.000 n=1+1)

name        old allocs/op  new allocs/op  delta
CRD/list-8       173 ± 0%         4 ± 0%   ~     (p=1.000 n=1+1)
CRD/get-8      34.4k ± 0%      0.2k ± 0%   ~     (p=1.000 n=1+1)
```

* Less complicated code; more complexity is handled by Kubernetes libraries/client-go codgen
* Kubernetes uses the pattern of importing client-go into their k/k repo, so it has prior art
* This will help us start fuzzing CRDs because we can use k8s libraries to do the fuzzing now that we have less custom code