@istio-testing 2020-06-08
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
lint_istio | 335420ff43a594290470df6a1305a7cc17d0c203 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/24511/lint_istio/13782) | `/test lint_istio`
gencheck_istio | 335420ff43a594290470df6a1305a7cc17d0c203 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/24511/gencheck_istio/12377) | `/test gencheck_istio`
unit-tests_istio | 335420ff43a594290470df6a1305a7cc17d0c203 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/24511/unit-tests_istio/14306) | `/test unit-tests_istio`
integ-pilot-k8s-tests_istio | 335420ff43a594290470df6a1305a7cc17d0c203 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/24511/integ-pilot-k8s-tests_istio/13889) | `/test integ-pilot-k8s-tests_istio`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->