@louiscryan 2020-06-11
I believe it was discussed in WG, can we record how we will support filesystem ingestion for ad-hoc/testing use-cases.