@howardjohn 2020-06-11
> I believe it was discussed in WG, can we record how we will support filesystem ingestion for ad-hoc/testing use-cases.

I think this is orthogonal to this change. This change is just replacing the current CRD reading code, which implements the IstioStore interface (basically just get/update/create on model.Config). Additionally, and unchanged in this PR, we have a filesystem reading code which implements the same IstioStore interface.